requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.7
sqlparse==0.5.3
tzdata==2025.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
